Add files via upload
This is a project im planning to run on a RPI4 with a 7 inch touchscreen. 
It will be used to gather average speeds of an inland vessel on the River Rhine. 
I want to save the data per trip into a CSV file, so I can load them up into excel.
There I will compare different trips with different variables for a highly accurate ETA. 

i have still a lot to do.